% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMDR.R
\name{classifyAllMDR}
\alias{classifyAllMDR}
\title{Classify the multi-drug resistance for multiple samples}
\usage{
classifyAllMDR(drug_resistance, antimicrobial_agents)
}
\arguments{
\item{drug_resistance}{A list with each row representing each
sample IDs and one of the columns is the string of
drugs that the sample resistant.}

\item{antimicrobial_agents}{The string of the column name specifying sample's antimicrobial agents.}
}
\value{
Returns a dataframe with each row name representing the sample ID and its
column representing the MDR category.
}
\description{
A function that classify the multi-drug resistances of all samples in the csv file
given the all drug resistances information from RGI prediction.
}
\examples{
# Example 2
# Using DrugResistance dataset available with package
dim(DrugResistance)

#classify the category of all samples' multi-drug resistance
resultsExample2 <- classifyAllMDR(drug_resistance = DrugResistance,
                               antimicrobial_agents = 'Antimicrobial.Agent')
# To obtain value from results
resultsExample2

}
\references{

}
