% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMDR.R
\name{classifyMDR}
\alias{classifyMDR}
\title{Classify the multi-drug resistance}
\usage{
classifyMDR(drug_resistance, sample_ID, antimicrobial_agents)
}
\arguments{
\item{drug_resistance}{A list with each key representing the sample ID and its
value is the string of the drugs that the sample resistant.}

\item{sample_ID}{The string of sample ID that used for classifying multi-drug resistance.}

\item{antimicrobial_agents}{The string of the column name specifying sample's antimicrobial agents.}
}
\value{
Returns the category of the given sample's multi-drug resistance.
}
\description{
A function that classify the multi-drug resistance given
the sample ID and drug resistance information from RGI prediction.
}
\examples{
# Example 1
# Using DrugResistance dataset available with package
dim(DrugResistance)
colnames(DrugResistance)
resultsExample1 <- classifyMDR(drug_resistance = DrugResistance,
                               sample_ID = "GFBCEDDN_00939",
                               antimicrobial_agents = "Antimicrobial.Agent")
# To obtain value from results
resultsExample1

}
\references{

}
