% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMDR.R
\name{classifyMDRfromRSI}
\alias{classifyMDRfromRSI}
\title{Classify the multi-drug resistance from RSI table}
\usage{
classifyMDRfromRSI(RSI_table, sample_ID, total_cat)
}
\arguments{
\item{RSI_table}{A list with each row representing the sample ID and each column
is the abbreviation of antimicrobial.
If the isolate is resistant to the agent, 'R' is in the cell.
Otherwise, 'S' is in the cell.}

\item{sample_ID}{The string of sample ID that used for classifying multi-drug resistance.}

\item{total_cat}{The integer number of total antimicrobial category of species.}
}
\value{
Returns the category of the given sample's multi-drug resistance.Example 5Using RSI_table dataset available with package

dim(RSI_table)
colnames(RSI_table)
resultsExample5 <- classifyMDRfromRSI(RSI_table = RSI_table,
sample_ID = "PA1381", total_cat = 8)To obtain value from results

resultsExample5
}
\description{
A function that classify the multi-drug resistance from RSI table.
}
\references{

}
