% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMDR.R
\name{countCategory}
\alias{countCategory}
\title{helper function
count <- countCategory(RSI_table = RSI_table, sample_ID = "PA1381")
count}
\usage{
countCategory(RSI_table, sample_ID)
}
\description{
helper function
count <- countCategory(RSI_table = RSI_table, sample_ID = "PA1381")
count
}
