% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DrugResistance}
\alias{DrugResistance}
\title{Drug Resistance Identified for Pseudomonas Aeruginosa From
Build in Canada Innovation Program Experiment.}
\format{
A matrix with columns:
\describe{
\item{Sample ID}{The sample ID of each isolate}
\item{ARO name}{The antimicrobial resistant gene predicted for each sample}
\item{Drug Class}{The name of the drug class that the sample resistant}
\item{Antimicrobial Agent}{The antimicrobial agent of each sample predicted from the whole genome sequence}
\item{Resistance Mechanism}{The resistance mechanismm that sample used to develop resistance}
}
}
\source{
PHAC’s National Microbiology Laboratory (NML)
https://www.canada.ca/en/public-health/programs/national-microbiology-laboratory.html
}
\usage{
DrugResistance
}
\description{
The experiment identifying the whole genome seqeunce of 1984 Pseudomonas aeruginosa isolates.
Subsequent data cleaning and RGI blast process obtain their ARO name, drug class,
antimicrobial agents, and resistant mechanism information conducted in 2021.
}
\examples{
\dontrun{
 DrugResistance
}
}
\keyword{datasets}
