% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDRCategoryPlot.R
\name{MDRPlot}
\alias{MDRPlot}
\title{Plotting the percentage of each MDR, XDR, PDR category.}
\usage{
MDRPlot(dataframe)
}
\arguments{
\item{dataframe}{A dataframe that contains all sample IDs and their MDR categories.}
}
\value{
Returns a histogram.
}
\description{
A function that takes in a list of dataframes and convert the antimicrobial
resistance level to three categories: MDR, XDR, and PDR. The barplot indicates
the percentage of MDR, XDR, and PDR isolates in the data.
}
\examples{
# Examples 1:
# plot the percentage
x <- classifyAllMDR(drug_resistance = DrugResistance[1:100,],
                               antimicrobial_agents = 'Antimicrobial.Agent')
MDRPlot(x)


}
