% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pca_data}
\alias{pca_data}
\title{WHONET 2019 software: \url{http://www.whonet.org/software.html}}
\format{
A matrix with columns:
\describe{
\item{Identification number}{The ID of each patient}
\item{Specimen number}{The ID of the specimen}
\item{Country}{The country of origin}
...
\item{AMP_ND10:CIP_EE}{Different antibiotics.}
}
\source{
AMR(for R) package on Github
https://github.com/msberends/AMR/blob/main/data/WHONET.rda
}
\usage{
pca_data
}
\description{
The experiment provided RSI information for each patient together with metadata.
}
\examples{
\dontrun{
 pca_data
}
}
\keyword{datasets}
