% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAMR.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA.}
\usage{
plotPCA(dataframe, new_data)
}
\arguments{
\item{dataframe}{A dataframe with rownames representing the sample ID and each column
is one of sample's antimicrobial features. If the value equals to 1,
the sample has this antimicrobial feature. If the value equals to 0,
the sample does not have this antimicrobial feature.}

\item{new_data}{The new data of dataframe must contain columns (variables) with the same
names and in the same order as the data used to compute PCA.}

\item{plot_type}{The type of plot need to be generated. For individual PCA plot,
using "individual". For visualizing relationship between variables, using
"variables". For predict new data and its relationship with old data, using
"predict".}
}
\value{
Returns the PCA plot.
}
\description{
A function that generate PCA and dimension reduction plots.
}
\examples{
Example 4
Using pca_data available with package
dim(pca_data)
resultsExample4 <- plotPCA(pca_data[c(2:10), c(2,9)], pca_data[c(11:13), c(2,9)])
resultsExample4

}
\references{

}
