% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictAMR.R
\name{predictAMR}
\alias{predictAMR}
\title{Predict AMR from PCA.}
\usage{
predictAMR(dataframe, new_data)
}
\arguments{
\item{dataframe}{A dataframe with rownames representing the sample ID and each column
is one of sample's antimicrobial features. If the value equals to 1,
the sample has this antimicrobial feature. If the value equals to 0,
the sample does not have this antimicrobial feature.}

\item{new_data}{The new data of dataframe must contain columns (variables) with the same
names and in the same order as the data used to compute PCA.}
}
\value{
Returns the PCA result of the new sample.
}
\description{
A function that predict multi-drug resistance by principle component analysis (PCA)
to show the new sample's in the original cluster.
}
\examples{
Example 3
Using pca_data available with package
dim(pca_data)
resultsExample3 <- predictAMR(pca_data[c(1:17), c(2,9,16)], pca_data[c(18:20), c(2,9,16)])
resultsExample3

}
\references{

}
