% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RSI_table}
\alias{RSI_table}
\title{Intensity of Antibiotic Resistance for Pseudomonas Aeruginosa From
RSI Experiment.}
\format{
A matrix with columns:
\describe{
\item{AMK}{Whether the isolate is susceptable to amikacin}
...
\item{TZP}{Whether the isolate is susceptable to tazobactam}
}
}
\source{
PHAC’s National Microbiology Laboratory (NML)
https://www.canada.ca/en/public-health/programs/national-microbiology-laboratory.html
}
\usage{
RSI_table
}
\description{
The experiment identifying the intensity of antibiotic resistance of Pseudomonas Aeruginosa conducted in 2020.
}
\examples{
\dontrun{
 RSI_table
}
}
\keyword{datasets}
