% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMDRClassifier.R
\name{runMDRClassifier}
\alias{runMDRClassifier}
\title{Launch Shiny App for MDRClassifier}
\usage{
runMDRClassifier()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for MDRClassifier.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

MDRClassifier::runMDRClassifier()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
