% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyMDR.R
\name{classifyAllMDRfromRSI}
\alias{classifyAllMDRfromRSI}
\title{Classify the multi-drug resistance for multiple samples}
\usage{
classifyAllMDRfromRSI(RSI_table)
}
\arguments{
\item{RSI_table}{A list with each row representing each
sample IDs and one of the columns is the string of
drugs that the sample resistant.}
}
\value{
Returns a dataframe with each row name representing the sample ID and its
column representing the MDR category.
}
\description{
A function that classify the multi-drug resistances of all samples in the csv file
given the all drug resistances information from RGI prediction.
}
\examples{
# Example 6
# Using DrugResistance dataset available with package

#classify the category of all samples' multi-drug resistance
resultsExample6 <- classifyAllMDRfromRSI(RSI_table = RSI_table)
# To obtain value from results
resultsExample6

}
\references{

}
