% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDRCategoryPlot.R
\name{MDRPlot}
\alias{MDRPlot}
\title{Plotting the percentage of each MDR, XDR, PDR category.}
\usage{
MDRPlot(dataframe)
}
\arguments{
\item{dataframe}{A dataframe that contains all sample IDs and their MDR categories.}
}
\value{
Returns a histogram.
}
\description{
A function that takes in a list of dataframes and convert the antimicrobial
resistance level to three categories: MDR, XDR, and PDR. The barplot indicates
the percentage of MDR, XDR, and PDR isolates in the data.
}
\examples{
# Examples 1:
# plot the percentage
x <- data.frame(Sample_ID = c("1","2","3","4","5","6"),
                Category = c("MDR", "PDR", "NULL", "XDR", "PDR", "NULL"))
MDRPlot(x)


}
